Refactor into separate commands (#20)
* Refactor into separate commands

* remove unnecesary import

* SF

* Updates README to point to new package home dir

* Add missing AddCommand()

* go mod tidy

* move log.SetFlags to main package
