+++ watch/watch.go
@@ -2,21 +2,41 @@ package watch
 
 import (
 	"fmt"
+	"io"
 	"os"
+	"sort"
 	"strings"
 
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-
+	"github.com/fatih/color"
+	"github.com/gosuri/uilive"
+	"github.com/pulumi/kubespy/pods"
 	"github.com/pulumi/pulumi-kubernetes/pkg/client"
+	"github.com/pulumi/pulumi-kubernetes/pkg/openapi"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/watch"
+	k8sWatch "k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/tools/clientcmd"
 
 	// Load auth plugins. Removing this will likely cause compilation error.
 	_ "k8s.io/client-go/plugin/pkg/client/auth"
+)
+
+const (
+	v1Endpoints = "v1/Endpoints"
+	v1Service   = "v1/Service"
+	prefix      = "\n       - "
+)
+
+var (
+	greenText    = color.New(color.FgGreen)
+	yellowText   = color.New(color.FgYellow)
+	cyanBoldText = color.New(color.FgCyan, color.Bold)
+	cyanText     = color.New(color.FgCyan)
+	redBoldText  = color.New(color.FgRed, color.Bold)
 )
 
 // Forever will watch a resource forever, emitting `watch.Event` until it is killed.
@@ -67,15 +87,113 @@ func Forever(apiVersion, kind, objID string) (<-chan watch.Event, error) {
 	return out, nil
 }
 
-func parseObjID(objID string) (string, string, error) {
-	split := strings.Split(objID, "/")
-	if l := len(split); l == 1 {
-		return "default", split[0], nil
-	} else if l == 2 {
-		return split[0], split[1], nil
+func PrintWatchTable(w *uilive.Writer, table map[string][]k8sWatch.Event) {
+	var svcType string
+	if events, hasSvc := table[v1Service]; hasSvc {
+		o := events[0].Object.(*unstructured.Unstructured)
+		watchEventHeader(w, events[0].Type, o)
+
+		svcTypeI, _ := openapi.Pluck(o.Object, "spec", "type")
+		var isString bool
+		svcType, isString = svcTypeI.(string)
+		if !isString {
+			svcType = "ClusterIP"
+		}
+
+		switch svcType {
+		case "ClusterIP":
+			if eps, hasEndpoints := table[v1Endpoints]; hasEndpoints && eps[0].Type != k8sWatch.Deleted {
+				printSuccessStatusEvent(w,
+					fmt.Sprintf("Successfully created Endpoints object '%s' to direct traffic to Pods",
+						cyanText.Sprint(o.GetName())))
+			} else {
+				printFailureStatusEvent(w, "Waiting for Endpoints object to be created, to direct traffic to Pods")
+			}
+
+			clusterIPI, _ := openapi.Pluck(o.Object, "spec", "clusterIP")
+			if clusterIP, isString := clusterIPI.(string); isString && len(clusterIP) > 0 {
+				printSuccessStatusEvent(w,
+					fmt.Sprintf("Successfully allocated a cluster-internal IP: %s",
+						cyanText.Sprint(o.GetName())))
+			} else {
+				printFailureStatusEvent(w, "Waiting for cluster-internal IP to be allocated")
+			}
+		case "LoadBalancer":
+			if eps, hasEndpoints := table[v1Endpoints]; hasEndpoints && eps[0].Type != k8sWatch.Deleted {
+				printSuccessStatusEvent(w,
+					fmt.Sprintf("Successfully created Endpoints object '%s' to direct traffic to Pods", o.GetName()))
+			} else {
+				printFailureStatusEvent(w, "Waiting for Endpoints object to be created, to direct traffic to Pods")
+			}
+
+			ingressesI, _ := openapi.Pluck(o.Object, "status", "loadBalancer", "ingress")
+			if ingresses, isMap := ingressesI.([]interface{}); isMap {
+				ips := []string{}
+				for _, ingressI := range ingresses {
+					if ingress, isMap := ingressI.(map[string]interface{}); isMap {
+						tmp := []string{}
+						ipI, _ := ingress["ip"]
+						ip, isString := ipI.(string)
+						if isString {
+							tmp = append(tmp, cyanText.Sprint(ip))
+						}
+
+						hostnameI, _ := ingress["hostname"]
+						hostname, isString := hostnameI.(string)
+						if isString {
+							tmp = append(tmp, cyanText.Sprint(hostname))
+						}
+
+						ips = append(ips, strings.Join(tmp, "/"))
+					}
+				}
+
+				if len(ips) > 0 {
+					sort.Strings(ips)
+					li := prefix + strings.Join(ips, prefix)
+					printSuccessStatusEvent(w, fmt.Sprintf("Service allocated the following IPs/hostnames:%s", li))
+				}
+			} else {
+				printFailureStatusEvent(w, "Waiting for public IP/host to be allocated")
+			}
+
+		case "ExternalName":
+			externalNameI, _ := openapi.Pluck(o.Object, "spec", "externalName")
+			if externalName, isString := externalNameI.(string); isString && len(externalName) > 0 {
+				printSuccessStatusEvent(w, fmt.Sprintf("Service proxying to %s", cyanText.Sprint(externalName)))
+			} else {
+				printFailureStatusEvent(w, "Service not given a URI to proxy to in `.spec.externalName`")
+			}
+		}
 	}
-	return "", "", fmt.Errorf(
-		"Object ID must be of the form <name> or <namespace>/<name>, got: %s", objID)
+
+	fmt.Fprintln(w)
+
+	if events, hasEPs := table[v1Endpoints]; hasEPs {
+		o := events[0].Object.(*unstructured.Unstructured)
+		watchEventHeader(w, events[0].Type, o)
+
+		ready := pods.GetReady(o)
+		sort.Strings(ready)
+		unready := pods.GetUnready(o)
+		sort.Strings(unready)
+		pods := append(ready, unready...)
+
+		if len(unready) > 0 {
+			li := prefix + strings.Join(pods, prefix)
+			printFailureStatusEvent(w, fmt.Sprintf("Directs traffic to the following live Pods:%s", li))
+		} else if len(ready) > 0 {
+			li := prefix + strings.Join(pods, prefix)
+			printSuccessStatusEvent(w, fmt.Sprintf("Directs traffic to the following live Pods:%s", li))
+		} else if len(unready) == 0 && len(ready) == 0 {
+			printFailureStatusEvent(w, fmt.Sprintf("Does not direct traffic to any Pods"))
+		}
+
+	} else if svcType != "ExternalName" {
+		fmt.Fprintln(w, "❌ Waiting for live Pods to be targeted by service")
+	}
+
+	w.Flush()
 }
 
 func makeClient() (discovery.CachedDiscoveryInterface, dynamic.ClientPool, error) {
@@ -109,4 +227,34 @@ func makeClient() (discovery.CachedDiscoveryInterface, dynamic.ClientPool, error
 	// apps, etc.)
 	pool := dynamic.NewClientPool(conf, mapper, pathresolver)
 	return discoCache, pool, nil
+}
+
+func parseObjID(objID string) (string, string, error) {
+	split := strings.Split(objID, "/")
+	if l := len(split); l == 1 {
+		return "default", split[0], nil
+	} else if l == 2 {
+		return split[0], split[1], nil
+	}
+	return "", "", fmt.Errorf(
+		"Object ID must be of the form <name> or <namespace>/<name>, got: %s", objID)
+}
+
+func printSuccessStatusEvent(w io.Writer, message string) {
+	fmt.Fprintf(w, "    ✅ %s\n", message)
+}
+
+func printFailureStatusEvent(w io.Writer, message string) {
+	fmt.Fprintf(w, "    ❌ %s\n", message)
+}
+
+func watchEventHeader(w io.Writer, eventType k8sWatch.EventType, o *unstructured.Unstructured) {
+	var eventTypeS string
+	if eventType == k8sWatch.Deleted {
+		eventTypeS = redBoldText.Sprint(eventType)
+	} else {
+		eventTypeS = greenText.Sprint(eventType)
+	}
+	apiType := cyanBoldText.Sprintf("%s/%s", o.GetAPIVersion(), o.GetKind())
+	fmt.Fprintf(w, "[%s %s]  %s/%s\n", eventTypeS, apiType, o.GetNamespace(), o.GetName())
 }
